pyproject.toml
[project]
name = "ddr-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "muradeliyev", email = "muradaliyev2229@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "camelot-py[cv]>=1.0.9",
    "einops>=0.8.1",
    "groq>=0.37.1",
    "ipykernel>=7.1.0",
    "langchain>=1.2.6",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-groq>=1.1.1",
    "pandas>=2.3.3",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.9",
    "pillow>=12.1.0",
    "protobuf==4.25.3",
    "python-dotenv>=1.2.1",
    "sentencepiece==0.2.1",
    "streamlit>=1.53.0",
    "timm>=1.0.24",
    "torch>=2.9.1",
    "transformers>=4.57.6",
    "watchdog>=6.0.0",
]

[project.scripts]
ddr-assistant = "ddr_assistant:main"

[build-system]
requires = ["uv_build>=0.9.13,<0.10.0"]
build-backend = "uv_build"
